+++ P9Tareas/MVVM/Converters/CodeToWeatherConverter.cs
@@ -11,7 +11,7 @@ namespace MauiWeather.Converters
      {
           public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
           {
-               var code = (float)value;
+               var code = (int)value;
 
                switch (code)
                {
